+++ plugins/11_TheWarWithin/localization/zhTW.lua
@@ -256,6 +256,7 @@ L['options_icons_no_harm_from_reading_desc'] = '顯示成就 {achievement:40632}
 L['no_harm_from_reading_note'] = nil
 L['another_you_4_note'] = nil
 
+L['concealed_contraband_note'] = nil
 L['memory_cache_note'] = '從附近 {spell:420847} from a nearby ({dot:Blue}) {object:Extractor Storage}. After kill the {npc:223908} to get the {item:223870} to open the {object:記憶寶箱}.'
 L['niffen_stash_note'] = '在橋的下面。'
 L['trapped_trove_note'] = nil